chore: re-organize summary modules and snapshots a little.
  Create a new module 'stake_distribution' to split a bit the code between pure rewards calculation and stake distribution.
