+++ examples/shared/src/store.rs
@@ -4,7 +4,7 @@ use amaru_ledger::{
         EpochTransitionProgress, HistoricalStores, ReadOnlyStore, Snapshot, Store, StoreError,
         TransactionalContext,
     },
-    summary::rewards::{Pots, RewardsSummary},
+    summary::{rewards::RewardsSummary, Pots},
 };
 use std::collections::BTreeSet;
 
@@ -40,9 +40,7 @@ impl ReadOnlyStore for MemoryStore {
         Ok(None)
     }
 
-    fn pots(
-        &self,
-    ) -> Result<amaru_ledger::summary::rewards::Pots, amaru_ledger::store::StoreError> {
+    fn pots(&self) -> Result<amaru_ledger::summary::Pots, amaru_ledger::store::StoreError> {
         Ok(Pots {
             fees: 0,
             treasury: 0,
@@ -312,4 +310,3 @@ impl HistoricalStores for MemoryStore {
         Ok(MemoryStore {})
     }
 }
-
